refactor(model-selector): drop V2 suffix and reclaim canonical name
- Move legacy `ModelSelector.tsx` to `ModelSelectorLegacy.tsx` and point
  its 6 page consumers at the new path.
- Move `ModelSelectorV2/` to `ModelSelector/`; rename the inner
  `ModelSelectorV2.tsx` / its test / logger context / describe block.
- Update the only external consumer (`ComponentLabModelSelectorSettings`)
  and the 13 i18n strings that referenced the V2 name.
