+++ src/renderer/src/components/ModelSelector/FreeTrialModelTag.tsx
