+++ src/renderer/src/components/ModelSelector/ModelTagChip.tsx
