+++ src/renderer/src/components/ModelSelector/__tests__/selection.test.ts
