+++ src/renderer/src/components/ModelSelector/__tests__/useModelListKeyboardNav.test.ts
