+++ src/renderer/src/components/ModelSelector/__tests__/utils.test.ts
