+++ src/renderer/src/components/ModelSelector/index.ts
@@ -1,6 +1,6 @@
 export type { ModelSelectorTag } from './filters'
 export { useModelTagFilter } from './filters'
-export { ModelSelector } from './ModelSelectorV2'
+export { ModelSelector } from './ModelSelector'
 export type {
   FlatListItem,
   ModelSelectorAlign,
