+++ src/renderer/src/components/ModelSelector/useModelListKeyboardNav.ts
