+++ src/renderer/src/components/ModelSelector/utils.ts
