+++ src/renderer/src/components/ModelSelectorLegacy.tsx
